requirement.txt
pyautogui==0.9.54
keyboard==0.13.5
ultralytics==8.3.99
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
torch>=2.0.0
torchvision>=0.15.0
